+++ src/org/camunda/latera/bss/connectors/MailSender.groovy
@@ -73,7 +73,7 @@ class MailSender {
     this.setFrom(user)
   }
 
-  MailSender getHost() {
+  String getHost() {
     return this.host
   }
 
@@ -83,7 +83,7 @@ class MailSender {
     return this
   }
 
-  MailSender getPort() {
+  Integer getPort() {
     return this.port
   }
 
@@ -97,7 +97,7 @@ class MailSender {
     return setPort(port.toInteger())
   }
 
-  MailSender getUser() {
+  String getUser() {
     return this.user
   }
 
@@ -135,10 +135,10 @@ class MailSender {
     return this
   }
 
-  MailSender addFile(CharSequence filename, Object datasource){
+  MailSender addFile(CharSequence name, Object datasource) {
     MimeBodyPart part = new MimeBodyPart()
     part.setDataHandler(new DataHandler(datasource, 'application/octet-stream'))
-    part.setFileName(filename.toString())
+    part.setFileName(MimeUtility.encodeText(name, 'UTF-8', null))
     multipart.addBodyPart(part)
     return this
   }
@@ -152,15 +152,20 @@ class MailSender {
     return this
   }
 
-  void send(){
+  Boolean send() {
+    Boolean result = true
     Transport transport = session.getTransport('smtp')
     transport.connect(host, port, user, password)
     try {
       message.setContent(multipart)
       transport.sendMessage(message, message.getAllRecipients())
-    }
-    finally {
+    } catch (Exception e) {
+      logger.error(e)
+      result = false
+    } finally {
       transport.close()
     }
+
+    return result
   }
 }